update response && gen json && set key in redis :money_with_wings:

+++ internals/domains/models/json.go
@@ -3,6 +3,6 @@ package models
 type JsonStructure struct {
 	Table    string                   `json:"table"`
 	DataKey  string                   `json:"key"`
-	Quantity int                      `json:"quantity"`
+	Quantity int64                    `json:"quantity"`
 	Datas    []map[string]interface{} `json:"datas"`
 }
